Accept mining fee for TransactionBuilder
Transaction builder is computing `mining_fee` for the transaction. User/clinet is not able to set the mining fee. Enable TransactionBuilder to accept mining fee and use it as the mining fee instead of computing the fee and make sure it compute the fee unless the mining fee is not provided.